fix: 1. user-project-list 404 2. check quotas of all projects when set domain quota 3. host.cpu shows negative number

+++ pkg/compute/models/hosts.go
@@ -2061,8 +2061,8 @@ type HostStat struct {
 	MemSize     int
 	MemReserved int
 	MemCmtbound float32
-	CpuCount    int8
-	CpuReserved int8
+	CpuCount    int
+	CpuReserved int
 	CpuCmtbound float32
 	StorageSize int
 }
